requirements.txt
alembic==1.7.5
aniso8601==9.0.1
black==21.11b1
cachelib==0.4.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
click==8.0.3
colorama==0.4.4
cryptography==36.0.0
Flask==2.0.2
Flask-HTTPAuth==4.5.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
httpauth==0.3
idna==3.3
importlib-metadata==4.8.2
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.3
jwt==1.3.1
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-enum==1.5.1
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
psycopg2==2.9.2
pycparser==2.21
PyJWT==2.3.0
python-decouple==3.5
pytz==2021.3
regex==2021.11.10
requests==2.26.0
six==1.16.0
SQLAlchemy==1.4.27
tomli==1.2.2
typing_extensions==4.0.0
urllib3==1.26.7
Werkzeug==2.0.2
zipp==3.6.0
